fix programs spilling over out of bounds
Some programs have a "program specified minimum size" that is
overriding the column width defined by WinTile. To combat this,
if a window spawns wider than anticipated, we will shift left by
however many pixels it overflows.  Likewise, we do the same thing
vertically if it spills over the bottom.

The side effect of this is that the window will underlap or overlap
the column directly to its left or below.